// course2/opi/lab3/Shooting-gallery-website/frontend/src/models/enums/ActionEnum.ts
enum ActionEnum{
    SET_USERNAME = "SET_USERNAME",
    SET_ACCESS_TOKEN = "SET_ACCESS_TOKEN",
    SET_LAST_POINT = "SET_LAST_POINT",
    SET_POINTS = "SET_POINTS",
    SET_POINT_COUNT = "SET_POINT_COUNT",
    SET_CURR_PAGE = "SET_CURR_PAGE",
    SET_POINT_STATE = "SET_POINT_STATE",
    SET_APP_STATE = "SET_APP_STATE",
    ADD_POPUP = "ADD_POPUP",
    REMOVE_POPUP = "REMOVE_POPUP",
    CLEAR_POPUPS = "CLEAR_POPUPS"
};

export default ActionEnum;